+++ internal/ratelimiter/ratelimiter.go
@@ -9,6 +9,7 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_RateLimiter_V1"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_RateLimiter"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/operation"
 	"github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter"
 	"github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter/config"
 )
@@ -66,12 +67,22 @@ func New(cc grpc.ClientConnInterface, options []config.Option) Client {
 	}
 }
 
+func (c *client) ctx(ctx context.Context) context.Context {
+	return operation.WithCancelAfter(
+		operation.WithTimeout(
+			ctx,
+			c.config.OperationParams().Timeout,
+		),
+		c.config.OperationParams().CancelAfter,
+	)
+}
+
 func (c *client) CreateResource(
 	ctx context.Context,
 	coordinationNodePath string,
 	resource ratelimiter.Resource,
 ) (err error) {
-	_, err = c.service.CreateResource(ctx, &Ydb_RateLimiter.CreateResourceRequest{
+	_, err = c.service.CreateResource(c.ctx(ctx), &Ydb_RateLimiter.CreateResourceRequest{
 		CoordinationNodePath: coordinationNodePath,
 		Resource: &Ydb_RateLimiter.Resource{
 			ResourcePath: resource.ResourcePath,
@@ -91,7 +102,7 @@ func (c *client) AlterResource(
 	coordinationNodePath string,
 	resource ratelimiter.Resource,
 ) (err error) {
-	_, err = c.service.AlterResource(ctx, &Ydb_RateLimiter.AlterResourceRequest{
+	_, err = c.service.AlterResource(c.ctx(ctx), &Ydb_RateLimiter.AlterResourceRequest{
 		CoordinationNodePath: coordinationNodePath,
 		Resource: &Ydb_RateLimiter.Resource{
 			ResourcePath: resource.ResourcePath,
@@ -111,7 +122,7 @@ func (c *client) DropResource(
 	coordinationNodePath string,
 	resourcePath string,
 ) (err error) {
-	_, err = c.service.DropResource(ctx, &Ydb_RateLimiter.DropResourceRequest{
+	_, err = c.service.DropResource(c.ctx(ctx), &Ydb_RateLimiter.DropResourceRequest{
 		OperationParams:      nil,
 		CoordinationNodePath: coordinationNodePath,
 		ResourcePath:         resourcePath,
@@ -129,7 +140,7 @@ func (c *client) ListResource(
 		response *Ydb_RateLimiter.ListResourcesResponse
 		result   Ydb_RateLimiter.ListResourcesResult
 	)
-	response, err = c.service.ListResources(ctx, &Ydb_RateLimiter.ListResourcesRequest{
+	response, err = c.service.ListResources(c.ctx(ctx), &Ydb_RateLimiter.ListResourcesRequest{
 		CoordinationNodePath: coordinationNodePath,
 		ResourcePath:         resourcePath,
 	})
@@ -152,7 +163,7 @@ func (c *client) DescribeResource(
 		response *Ydb_RateLimiter.DescribeResourceResponse
 		result   Ydb_RateLimiter.DescribeResourceResult
 	)
-	response, err = c.service.DescribeResource(ctx, &Ydb_RateLimiter.DescribeResourceRequest{
+	response, err = c.service.DescribeResource(c.ctx(ctx), &Ydb_RateLimiter.DescribeResourceRequest{
 		CoordinationNodePath: coordinationNodePath,
 		ResourcePath:         resourcePath,
 	})
@@ -201,6 +212,6 @@ func (c *client) AcquireResource(
 			Units:                &Ydb_RateLimiter.AcquireResourceRequest_Required{Required: amount},
 		}
 	}
-	_, err = c.service.AcquireResource(ctx, &request)
+	_, err = c.service.AcquireResource(c.ctx(ctx), &request)
 	return
 }